Upgrade to only support v2 bash_completion.
This eliminates some cruft for supporting bash_completion v1, which is
really unnecessary at this point. Even servers used within all
infrastructure I manage are modern and support v2, like CentOS 7.

Note that this work was done in response to debugging why git
completions were not working. In the end, it had to do with the
installation of GitHub's hub utility: github/hub#2684. Ultimatelty,
hub was removed until that is fixed.