Fix issue with brew reading stdin in while loop.
This has never happened before, yet now it seems like homebrew is
reading from stdin during a while loop. It was reading the first line,
installing, then mid-install it would dump out the rest of the file,
finish the install, then carry on to the next block of logic. This kept
me busy all day.