Store BrewFile and lock file from previous build.
This is just a snapshot from the last MBP, in added in case there were
critical apps that were overlooked. This was generated
using `brew bundle dump`, for reference.

This whole process may see a complete rewrite, with either smaller
BrewFiles or a completely different build system, like with Nix.