Add README.md note about v2.0.0.
Anything beyond this point begins to leverage Home Manager with Nix
for all tooling and configuration.

There will still be some reliance on Homebrew, especially for Casks,
but an effort is being made to completely migrate away. This README
change will leave this note for users who may have been using the old
version and may wonder why the dotfiles no longer work as expected.

This commit will be merged into the feature/nix-home-manager branch to
reflect the state once its in master.