Migrate cURL to Home Manager.
This is an example of migrating a package that was being managed by
Homebrew in these dotfiles to being managed by Home Manager via Nix.

This one is easy because there are no real configs required to manage
it, but it does get manually added to the PATH, and have a custom
install script.

Other migrations will require things like managing configs through
Home Manager as well, or telling Home Manager to load a specific config
and then symlinking it itself, but these dotfiles have never managed a
custom dotfile config for cURL.