Migrate some other packages.
Most of the first packages that will get migrated over will be ones that
do not have personally managed dotfile configs for them. Additionally,
the ones being migrated over will be ones that have additional steps
like needing to add their installations to the PATH, which should all
be done by Nix in the backend.