Modify PATH building again.
The PATH was getting mangled in JetBrains, which was yet another
place that was rebuilding PATH on MacOS. This change means that
PATH will be rebuilt every session, identically, regardless of it
being a login shell or not, or being within TMUX, or being within
JetBrains.

Note that this solution effectively restores a much earlier iteration.