Re-introduce Brew compat from Nix. #25
This ensures compatibility to merge all the work
from the Nix Home Manager branch back to master.

In the end, it may not be too bad to simply
support both systems if they are live. Only
the paths and some environment variables will
need to account for this.

The next commit will add many of the packages
that were migrated to Nix Home Manager.