Add trimmed down brew packages. #25
All of the packages commented out with double brackets are
likely still necessary, but for the meantime, the dotfiles
are attempting to trim down on as much as possible for an
initial install.

Eventually even the functions, aliases, and various changes
will be audited to ensure only the minimal installation
works, to the point that the dotfiles files themselves
can simply exist.