Fix homebrew paths. #25
Now that Homebrew no longer loads in default path locations, but its own
within /opt/homebrew, the creation of environment variables needs to
occur at minimum after the paths have been added. Paths on non-M1 MBPs
never needed to be set for Homebrew by this point.

This also restores the coreutils utilities to the PATH if they exist,
which is especially important for readpath and ls.