Update dotfiles install script for brew. #25
This updates to the latest recommended install script, that no longer
depends on ruby.

This also adds Rosetta installation if available; it will only be
available on Apple Silicon MBPs.

This also does a brew installation of additional utilities, and ensures
that the path used is interoperable between M1 and non-M1 MBPs.