Allow bashrc to bootstrap from multiple paths.
This simply checks if the current environment leverages some of the
available iCloud shell functionalities built into these dotfiles, and if
so, will load from that path. The original path of ~/.dotfiles is still
relevant and will be loaded if it exists and the iCloud one does not.

This is just to keep my personal home directory clean, as it's unlikely
anyone besides me will ever use this iCloud functionality.