Fix viminfo issue.
A new .viminfo file was being created in the home directory constantly.
At first I thought it was due to the condition check, which is likely
faulty, but after some research it was discovered that the viminfo file
is not allowed to be a symlink due to a security reason and thus by
design the viminfo file setting will never be respected if it is a
symlink. Therefore, the fix was to nest it into a directory and have
that directory symlinked instead.