Simplify Homebrew path additions. #26
Note that prior to this the post-installation step for Homebrew was
never run (brew shellenv), which is responsible for adding relevant
paths for bin and sbin, as well as exporting several HOMEBREW_
environment variables.

This is also why the sbin path from homebrew was removed. Note that
technically this script should just run the shellenv command to get all
of the homebrew paths added (both bin and sbin), but it was decided that
maintaining the old /usr/local/bin and then having just the
/opt/homebrew/bin path, that the brew shellenv command could then just
be executed without knowing the absolute path. This maintains
compatibility on both older Intel Macs and Apple silicon Macs.