Whitelist path in /etc/shells for chsh . #26
This should fix the issue where the new bash shell is not being set to
the path provided by Homebrew. Prior to Sequoia at least, I do not
recall having this issue with changing the shell programmatically.
Their documentation does suggest that the path needs to be whitelisted
in /etc/shells to support changing the shell, though, so this fix does
that.

This also silences the zsh warning when using a shell that is not the
default (bash). It is probably unnecessary after whitelisting and then
setting the default to bash, but this will continue to keep things quiet
if ever the default shell is zsh and one wants to just dabble with bash
momentarily.